chore(skills-presentation-fragment): lint fix unused element
