--- site/course-description-fragment/registration-popup/index.ts
@@ -1,1 +1,0 @@
-export * from './registration-popup.component'
