+++ site/course-description-fragment/src/sign-up-form/index.ts
@@ -1,0 +1,1 @@
+export * from './sign-up-form.component'
