+++ site/skills-presentation-fragment/src/skills-presentation.component.tsx
@@ -5,7 +5,6 @@ import { Box }    from '@ui/layout'
 import { Column } from '@ui/layout'
 import { Layout } from '@ui/layout'
 import { Row }    from '@ui/layout'
-import { Text }   from '@ui/text'
 
 import { Item }   from './item.component'
